// Smart-Student-React/src/interfaces/General/GendersCivStatuNacLevEducRelati.ts
import { SelectedInterface } from "../SelectedInterface"

export interface GendersCivStatuNacLevEducRelati {
    civilStatus: SelectedInterface<number>[];
    nationality: SelectedInterface<number>[];
    educationalLevel: SelectedInterface<number>[];
    relationship: SelectedInterface<number>[];
    professions: SelectedInterface<number>[];
    genders: SelectedInterface<number>[];
    grades: SelectedInterface<number>[];
    levels: SelectedInterface<number>[];
}